pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lottery-generator"
version = "0.1.0"
description = "Gerador de apostas de loteria em Python, com estratégias configuráveis e filtros."
authors = [
    { name = "Matheus Beiruth Miranda dos Santos", email = "matheusbeiruth10@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["generator"]

[tool.pytest.ini_options]
pythonpath = ["."]
